[Design] Move UrlMatchingTree to shared source
